feat: type에 따른 book  디테일 페이지 분리
